Move test for default components to contrib
